All tests compile in all configurations
git-svn-id: svn://136.177.114.72/svn_GW/IPhreeqc/trunk@4357 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ IPhreeqcLib.cpp
@@ -154,7 +154,7 @@ GetComponentCount(int id)
 	IPhreeqc* IPhreeqcPtr = IPhreeqcLib::GetInstance(id);
 	if (IPhreeqcPtr)
 	{
-		return IPhreeqcPtr->ListComponents().size();
+		return (int)IPhreeqcPtr->ListComponents().size();
 	}
 	return IPQ_BADINSTANCE;
 }
